+++ bootique-cayenne-test/src/test/java/io/bootique/cayenne/test/CayenneTestDataManagerTxIT.java
@@ -4,11 +4,8 @@ import io.bootique.BQRuntime;
 import io.bootique.cayenne.test.persistence.Table1;
 import io.bootique.jdbc.test.Table;
 import io.bootique.test.junit.BQTestFactory;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
 
 public class CayenneTestDataManagerTxIT {
 
@@ -16,8 +13,6 @@ public class CayenneTestDataManagerTxIT {
     public BQTestFactory testFactory = new BQTestFactory();
 
     @Test
-    @Ignore
-    // TODO: unignore when #44 is fixed by upgrading to Cayenne 4.0.B2
     public void testDataSourceDoesNotAutocommit() {
 
         BQRuntime runtime = testFactory.app("-c", "classpath:config-noautocommit.yml")
@@ -31,8 +26,8 @@ public class CayenneTestDataManagerTxIT {
 
         Table t1 = dataManager.getTable(Table1.class);
 
-        assertEquals(0, t1.getRowCount());
+        t1.matcher().assertNoMatches();
         t1.insert(1, 2, 3);
-        assertEquals(1, t1.getRowCount());
+        t1.matcher().assertOneMatch();
     }
 }